[CHORE #160409504]- Create questions and move question classes
