+++ app/views/views.py
@@ -13,94 +13,6 @@ validate = FieldValidation()
 question_blueprint = Blueprint("question_blueprint", __name__)
 
 
-class PostQuestion(MethodView):
-    """class for posting new question"""
-
-    @jwt_required
-    def post(self):
-        try:
-            data = request.get_json()
-            search_keys = ("title", "question")
-
-            if all(key in data.keys() for key in search_keys):
-                now = datetime.datetime.now()
-
-                loggedin_user = get_jwt_identity()
-                user = get_user_by_username(user_name=loggedin_user["username"], password=loggedin_user["password"])
-
-                qstn_owner = user["username"]
-                title = data.get("title").strip()
-                question = data.get("question").strip()
-                date = now.strftime("%Y-%m-%d %H:%M")
-
-                validation = validate.validate_question(title, question)
-
-                if validation:
-                    return validation
-
-                does_qstn_exist = is_question_exist(question)
-                if does_qstn_exist:
-                    return jsonify({"message": "Question already exists, check it out for an answer"}), 409
-
-                post_new_question(title=title, question=question, qstn_owner=qstn_owner, date=date)
-
-                new_question = Question(title=title, question=question, qstn_owner=qstn_owner, date=date)
-                return jsonify({"New Question Posted": new_question.__dict__}), 201
-            return jsonify({"message": "a 'key(s)' is missing in your question body"}), 400
-        except:
-            return jsonify({"message": "All fields are required"}), 400
-
-
-class FetchAllQuestions(MethodView):
-    """Class to fetch all questions posted"""
-
-    @jwt_required
-    def get(self):
-        all_questions = get_all_questions()
-        if all_questions:
-            return jsonify({"All Questions": all_questions}), 200
-        return jsonify({"message": "No questions posted yet"}), 404
-
-
-class FetchSingleQuestion(MethodView):
-    """class to get single question"""
-
-    @jwt_required
-    def get(self, qstn_id):
-        try:
-            id_validation = validate.validate_entered_id(qstn_id)
-            if id_validation:
-                return id_validation
-
-            question_details = get_single_question(qstn_id=qstn_id)
-            all_answers = get_all_answers_to_question(qstn_id=qstn_id)
-            if question_details:
-                return jsonify({"Question Details": question_details, "Answers": all_answers}), 200
-            return jsonify({"message": "Question does not exist"}), 404
-        except:
-            return jsonify({"message": "Check your url and try again"}), 400
-
-
-class DeleteQuestion(MethodView):
-    """class to get single question"""
-
-    @jwt_required
-    def delete(self, qstn_id):
-        try:
-            # if type(qstn_id) is not int:
-            #     return jsonify({"message": "Question id should be an integer"}), 400
-            id_validation = validate.validate_entered_id(qstn_id)
-            if id_validation:
-                return id_validation
-            loggedin_user = get_jwt_identity()
-
-            question_details = get_single_question(qstn_id=qstn_id)
-            if question_details:
-                delete_question(qstn_id, loggedin_user["username"])
-                return jsonify({"message": "Question successfully deleted"}), 200
-            return jsonify({"message": "Question does not exist"}), 404
-        except:
-            return jsonify({"message": "Check your url and try again"}), 400
 
 
 class PostAnswerToQuestion(MethodView):
